Implement creola::core::FuncCall::simplify() method

+++ creola-core/common.hpp
@@ -25,7 +25,7 @@ using HashMap = std::unordered_map<Key, Value>;
 template<typename Key, typename Value>
 using Dict = std::map<Key, Value>;
 
-using f64 = long double;
+using f64 = double;
 using i64 = std::int64_t;
 using usize = size_t;
